Create Global_Change_Connect.github.io
This is a websit created for Global Change as a result of anthropogenic activities and also compendium of articles investigations on anthropogenic impact on the environment. The websit will also gather data of different paramenter and models for predictions.Latest invests and news on environmental and climate issues will apso be published.